+++ lib/mayaUsd/fileio/utils/CMakeLists.txt
@@ -5,6 +5,7 @@ target_sources(${PROJECT_NAME}
     PRIVATE
         adaptor.cpp
         meshReadUtils.cpp
+        meshWriteUtils.cpp
         readUtil.cpp
         roundTripUtil.cpp
         shadingUtil.cpp
@@ -16,6 +17,7 @@ target_sources(${PROJECT_NAME}
 set(headers
     adaptor.h
     meshReadUtils.h
+    meshWriteUtils.h
     readUtil.h
     roundTripUtil.h
     shadingUtil.h
